\name{impute}
\alias{impute}
\title{Impute}
\usage{
  impute(iterations = 1, options = "", ...)
}
\arguments{
  \item{iterations}{}

  \item{options}{}

  \item{...}{}
}
\value{
  NULL
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Jochen Kruppa
}

