\name{predict.rjungle}
\alias{predict.rjungle}
\title{predict}
\usage{
  predict.rjungle(object, data, ...)
}
\arguments{
  \item{object}{}

  \item{data}{}

  \item{...}{}
}
\value{
  NULL
}
\description{
  .. content for \description{} (no empty lines) ..
}
\details{
  .. content for \details{} ..
}
\author{
  Jochen Kruppa
}

