\name{Rjungle-package}
\alias{Rjungle-package}
\alias{Rjungle}
\docType{package}
\title{
Rjungle
}
\description{
Performs RandomJungle from within R and provides functions to access the results.
}
\details{
\tabular{ll}{
Package: \tab Rjungle\cr
Type: \tab Package\cr
Version: \tab 0.1-3\cr
Date: \tab 2010-09-07\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Daniel Schwarz with modifications by Andreas Bender
}
\references{
\url{www.randomjungle.org}
}
\keyword{ package }
\seealso{
\code{\link{rjungle}}, \code{\link{importance}}, \code{\link{confusion}}
}

